Add files via upload
I created the shell for all of the functions, added them to my index.js file but had some issues on .add vs .post, some issues actually uploading and connecting to the API and  also some issues understanding where to put what variable and why it goes where it goes. .then() I understand is the promise and the concept but again I just am a little confused on what to name things and where to implement them. I created the shell without referencing my notes but when I was filling out the function i didn't have the correct notes to follow and I didn't want to copy and paste so for me I felt that trying to understand the placement and memorizing the shell of the function was progress but I still have questions.